# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Простой файлообменник с ограниченным сроком жизни файлов",
        "title": "FileShare API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/batch/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batch"
                ],
                "summary": "Получить список файлов батча",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID батча",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/httpapi.DownloadLink"
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/batch/{id}/zip": {
            "get": {
                "description": "Архивирует все ещё не истёкшие файлы батча в zip и стримит его в ответ",
                "produces": [
                    "application/zip"
                ],
                "tags": [
                    "batch"
                ],
                "summary": "Скачать все файлы батча архивом",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID батча",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "batch not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/files/{id}": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Скачать файл",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID файла",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "410": {
                        "description": "expired",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/options": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Получить доступные варианты срока хранения",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/httpapi.ExpiryOptions"
                            }
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "Загружает один или несколько файлов и возвращает ссылку на батч",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Загрузить файлы",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Файлы для загрузки",
                        "name": "myFile",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Срок хранения (20s, 1d, 3d, 7d)",
                        "name": "expires",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httpapi.DownloadLink": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "httpapi.ExpiryOptions": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                }
            }
        }
    }
}